requirements.txt
absl-py==2.0.0
addict==2.4.0
aiohttp
aiosignal==1.3.1
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
arrow==1.3.0
asgiref==3.7.2
async-lru==2.0.4
attrs==23.1.0
babel==2.14.0
beautifulsoup4==4.12.2
black
bleach==6.1.0
blinker==1.8.1
click==8.1.7
contourpy==1.2.0
einops==0.7.0
flasgger==0.9.7.1
flask==3.0.3
flask-sqlalchemy==3.1.1
gitpython
gunicorn==21.2.0
h5py==3.10.0
huggingface-hub==0.20.1
ipykernel==6.27.1
ipython==8.19.0
ipywidgets==8.1.1
jedi==0.19.1
jinja2
joblib==1.3.2
jsonschema==4.20.0
kornia==0.7.1
lightning-utilities==0.10.0
matplotlib==3.8.2
minio~=7.2.10
numpy~=1.26.2
opencv-python==4.8.1.78
pandas==2.1.4
protobuf==4.23.4
pytorch-lightning
pymysql==1.1.1
requests~=2.31.0
scikit-image
scikit-learn
scipy==1.11.4
sqlalchemy==2.0.30
tensorboardx==2.6.2.2
tqdm==4.66.3
thop~=0.1.1-2209072238
wandb==0.16.2
pyiqa==0.1.4
Flask-PyMongo
redis~=5.2.0
torch~=2.1.0
torchvision~=0.16.0
pillow~=10.1.0
werkzeug~=3.0.1
pyyaml~=6.0.1
timm~=0.6.13
yacs~=0.1.8
setuptools~=59.5.0
ema-pytorch~=0.3.1
